move RottenReviewModal into its own component
